+++ packages/terra-search-field/src/terra-dev-site/test/search-field/SearchFieldEnter.test.jsx
@@ -5,7 +5,6 @@ import SearchField from '../../../SearchField';
 const locale = document.getElementsByTagName('html')[0].getAttribute('lang');
 
 class AutoSearchDisabledSearchField extends React.Component {
-
   constructor(props) {
     super(props);
 
@@ -41,7 +40,6 @@ class AutoSearchDisabledSearchField extends React.Component {
       </Base>
     );
   }
-
 }
 
 export default AutoSearchDisabledSearchField;