+++ packages/terra-search-field/CHANGELOG.md
@@ -9,6 +9,11 @@ Unreleased
 ### Removed
 * Removed a duplicate test page for Search Fields with Placeholders
 
+2.15.0 - (June 22, 2018)
+------------------
+### Changed
+* Moved terra-base from dependency to peerDependency
+
 2.14.0 - (June 12, 2018)
 ------------------
 ### Changed